[Backend] REST endpoints implement #7

+++ src/main/java/hu/dominikvaradi/sociallybackend/flows/message/domain/dto/MessageResponseDto.java
@@ -19,5 +19,5 @@ public class MessageResponseDto {
 	private UUID conversationId;
 	private String content;
 	private LocalDateTime created;
-	EnumMap<Reaction, Integer> reactionsCount;
+	EnumMap<Reaction, Long> reactionsCount;
 }
